+++ packages/mdx-remote/.eslintrc.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: ['custom/library'],
+  rules: {
+    'import/no-named-as-default': 'off',
+    'import/no-named-as-default-member': 'off',
+  },
+};